Refactored model
