+++ src/main/scala/LongBoardsApp.scala
@@ -1,6 +1,5 @@
 import akka.actor.{ActorSystem}
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.StatusCodes
 import akka.stream.SystemMaterializer
 import akka.http.scaladsl.server.Directives._
