+++ src/main/scala/model/LongBoard/LongBoard.scala
@@ -1,0 +1,3 @@
+package model.LongBoard
+
+case class LongBoard(_id: String, mark: String, model: String, quantity: Int)
